約数リスト (12 -> [1, 2, 3, 4, 6, 12])

+++ 0divisorList.java
@@ -1,10 +1,10 @@
-    public static List<Integer> divisorList(int n) {
-        List<Integer> res = new ArrayList<Integer>();
-        for (int i = 1; i * i <= n; i++) {
+    public static List<Long> divisorList(long n) {
+        List<Long> res = new ArrayList<Long>();
+        for (long i = 1; i * i <= n; i++) {
             if (n % i == 0) {
                 res.add(i);
                 
-                int j = n / i;
+                long j = n / i;
                 if (j != i) {
                     res.add(j);
                 }
